requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
ghidra-stubs==11.4
idna==3.11
requests==2.32.5
types-requests==2.32.4.20250913
urllib3==2.5.0
